+++ apps/expo-go/ios/Exponent/Versioned/Core/Modules/ExpoGoExpoUpdatesModule.swift
@@ -38,7 +38,7 @@ final class ExpoGoExpoUpdatesModule: Module {
 
       let releaseChannel = config?.releaseChannel
       let channel = config?.requestHeaders["expo-channel-name"] ?? ""
-      let runtimeVersion = config?.runtimeVersionRaw ?? ""
+      let runtimeVersion = config?.runtimeVersion ?? ""
       let checkAutomatically = config?.checkOnLaunch.asString ?? CheckAutomaticallyConfig.Always.asString
       let isMissingRuntimeVersion = false