+++ packages/@expo/config-plugins/src/android/__tests__/Updates-test.ts
@@ -42,6 +42,9 @@ describe('Android Updates config', () => {
     const config: ExpoConfig = {
       name: 'foo',
       sdkVersion: '37.0.0',
+      runtimeVersion: {
+        policy: 'sdkVersion',
+      },
       slug: 'my-app',
       owner: 'owner',
       updates: {
@@ -79,8 +82,7 @@ describe('Android Updates config', () => {
     const sdkVersion = mainApplication['meta-data'].filter(
       (e) => e.$['android:name'] === 'expo.modules.updates.EXPO_SDK_VERSION'
     );
-    expect(sdkVersion).toHaveLength(1);
-    expect(sdkVersion[0].$['android:value']).toMatch('37.0.0');
+    expect(sdkVersion).toHaveLength(0);
 
     const enabled = mainApplication['meta-data'].filter(
       (e) => e.$['android:name'] === 'expo.modules.updates.ENABLED'
@@ -125,11 +127,11 @@ describe('Android Updates config', () => {
       '{"expo-channel-name":"test","testheader":"test"}'
     );
 
-    // For this config, runtime version should not be defined, so check that it does not appear in the manifest
     const runtimeVersion = mainApplication['meta-data']?.filter(
       (e) => e.$['android:name'] === 'expo.modules.updates.EXPO_RUNTIME_VERSION'
     );
-    expect(runtimeVersion).toHaveLength(0);
+    expect(runtimeVersion).toHaveLength(1);
+    expect(runtimeVersion[0].$['android:value']).toMatch('@string/expo_runtime_version');
   });
 
   describe(Updates.ensureBuildGradleContainsConfigurationScript, () => {