+++ packages/expo-dev-menu/ios/EXDevMenuAppInfo.m
@@ -17,7 +17,6 @@
 
   NSString *appIcon = [EXDevMenuAppInfo getAppIcon];
   NSString *runtimeVersion = [EXDevMenuAppInfo getUpdatesConfigForKey:@"EXUpdatesRuntimeVersion"];
-  NSString *sdkVersion = [EXDevMenuAppInfo getUpdatesConfigForKey:@"EXUpdatesSDKVersion"];
   NSString *appVersion = [EXDevMenuAppInfo getFormattedAppVersion];
   NSString *appName = [[NSBundle mainBundle] objectForInfoDictionaryKey: @"CFBundleDisplayName"] ?: [[NSBundle mainBundle] objectForInfoDictionaryKey: @"CFBundleExecutable"];
 
@@ -44,7 +43,6 @@
   appInfo[@"appIcon"] = appIcon;
   appInfo[@"appVersion"] = appVersion;
   appInfo[@"runtimeVersion"] = runtimeVersion;
-  appInfo[@"sdkVersion"] = sdkVersion;
   appInfo[@"hostUrl"] = hostUrl;
   appInfo[@"engine"] = engine;